# Speed up component lifecycle